membuat offer bisa di create dengan mengambil id_user dari jwt

+++ internal/entity/offers.go
@@ -2,7 +2,7 @@ package entity
 
 type Offer struct {
 	IDOffer     int64  `json:"id_offer" gorm:"autoIncrement;"`
-	IDUser      int64  `json:"id_user"`
+	IDUser      int    `json:"id_user"`
 	Email       string `json:"email"`
 	NameEvent   string `json:"name_event"`
 	Description string `json:"description"`
